fix: focus doesn't go back to slide right after resizing
